% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plot.R
\name{bootSigsPlot}
\alias{bootSigsPlot}
\title{Plot signature contributions with 95 \% CI}
\usage{
bootSigsPlot(mutSigsSummary)
}
\arguments{
\item{mutSigsSummary}{data frame of bootstrapping output in the format from mm_fit_signatures}
}
\value{
plots showing mutational signature estimates with 95 \% CI by sample
}
\description{
Plot signature contributions with 95 \% CI
}
