% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_fit_signatures}
\alias{bootstrap_fit_signatures}
\title{Bootstrapping function for mmsig}
\usage{
bootstrap_fit_signatures(
  samples.muts,
  consigts.defn,
  sigt.profs,
  cos_sim_threshold,
  force_include,
  iterations = 1000
)
}
\arguments{
\item{samples.muts}{96 classes mutational profile for all samples}

\item{consigts.defn}{96 classes mutational profile for mutational signature reference}

\item{sigt.profs}{which signatures to fit for each sample}

\item{cos_sim_threshold}{cosine similarity threshold to remove signatures}

\item{force_include}{vector with the names of signatures to always keep in the final profile of every sample}

\item{iterations}{number of mutational profiles to draw from the multinomial distribution of each sample}
}
\value{
For each mutational signature, returns its relative contribution as point estimate and bootstrapping mean with 95 \% CI
}
\description{
Bootstrapping function for mmsig
}
