% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_signatures.R
\name{em_signatures}
\alias{em_signatures}
\title{Expectation maximization algorithm to estimate the signature contribution}
\usage{
em_signatures(sigts.defn, mut.freqs, max.iter, dbg)
}
\arguments{
\item{sigts.defn}{mutational signature definitions}

\item{mut.freqs}{mutational profile of sample}

\item{max.iter}{maximum number of iterations}

\item{dbg}{boolean whether to print or not}
}
\value{
estimated mutational signature contributions
}
\description{
Expectation maximization algorithm to estimate the signature contribution
}
