% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_signatures.R
\name{fit_signatures}
\alias{fit_signatures}
\title{Signature fitting function for mmsig}
\usage{
fit_signatures(
  samples.muts,
  consigts.defn,
  sigt.profs,
  cos_sim_threshold,
  force_include,
  dbg = dbg
)
}
\arguments{
\item{samples.muts}{96 classes mutational profile for all samples}

\item{consigts.defn}{96 classes mutational profile for mutational signature reference}

\item{sigt.profs}{which signatures to fit for each sample}

\item{cos_sim_threshold}{cosine similarity threshold to remove signatures}

\item{force_include}{vector with the names of signatures to always keep in the final profile of every sample}

\item{dbg}{boolean whether to print or not}
}
\value{
estimated mutational signature profile of all samples
}
\description{
Signature fitting function for mmsig
}
