% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_5_col.R
\docType{data}
\name{mm_5_col}
\alias{mm_5_col}
\title{Example data in 5-columns format}
\format{
A data frame with 445813 rows and 5 variables:
\describe{
\item{sample}{sample ID}
\item{chr}{chromosome}
\item{pos}{position on the hg19 reference genome}
\item{ref}{nucleotide in the hg19 reference genome}
\item{alt}{alternative nucleotide in the sample}
...
}
}
\source{
\url{https://www.nature.com/articles/s41467-020-15740-9}
}
\usage{
mm_5_col
}
\description{
Complete mutational catalogs from whole genome sequencing of 82 multiple myeloma samples
}
\keyword{datasets}
