% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_96_classes.R
\docType{data}
\name{mm_96_classes}
\alias{mm_96_classes}
\title{Example data in 96 classes format}
\format{
Mutational count matrix with 96 rows and 82 variables:
\describe{
\item{row names}{mutational class}
\item{column names}{sample ID}
...
}
}
\source{
\url{https://www.nature.com/articles/s41467-020-15740-9}
}
\usage{
mm_96_classes
}
\description{
Complete mutational catalogs from whole genome sequencing of 82 multiple myeloma samples
}
\keyword{datasets}
