% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mm_fit_signatures}
\alias{mm_fit_signatures}
\title{Mutational signature fitting with mmsig}
\usage{
mm_fit_signatures(
  muts.input,
  sig.input,
  input.format = "vcf",
  sample.sigt.profs = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  strandbias = FALSE,
  refcheck = TRUE,
  cos_sim_threshold = 0.01,
  force_include = c("SBS1", "SBS5"),
  dbg = FALSE
)
}
\arguments{
\item{muts.input}{mutation input data frame as specified under input.format}

\item{sig.input}{mutational signature reference with mutational classes as rows and signature exposures as columns: Substitution.Type, Trinucleotide, signature.1, ... signature.n}

\item{input.format}{vcf: five column vcf-like data frame with the following columns: sample, chr (e.g. chr1), pos, ref, alt. classes: samples as columns and the 96 mutational classes as rows}

\item{sample.sigt.profs}{NULL = use all signatures provided in the reference. Optionally provide list with signatures to consider for each sample}

\item{bootstrap}{TRUE/FALSE for whether bootstrapping is to be performed}

\item{iterations}{number of bootstrapping iterations to perform (only if bootstrap == TRUE)}

\item{strandbias}{TRUE/FALSE for whether transcriptional strand bias should be tested for (only for vcf-like input format)}

\item{refcheck}{check that input mutational catalog (if vcf-format) is aligned to hg19}

\item{cos_sim_threshold}{cosine similarity threshold below which signatures are removed from the final profile}

\item{force_include}{vector with the names of signatures to always keep in the final profile of every sample}

\item{dbg}{FALSE = silent; TRUE = verbose}
}
\value{
mutational signature fitting results for all samples
}
\description{
Mutational signature fitting with mmsig
}
