% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_plot.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plotting function for mmsig package}
\usage{
plot_signatures(
  sig,
  sig_order = c("SBS1", "SBS2", "SBS13", "SBS5", "SBS8", "SBS9", "SBS18", "SBS-MM1",
    "SBS35"),
  samples = FALSE
)
}
\arguments{
\item{sig}{data frame of mutational signature estimates in the format from mm_fit_signatures}

\item{sig_order}{order of signatures to plot in stacked bar chart, from top to bottom}

\item{samples}{boolean whether or not to plot sample names on the x axis}
}
\value{
relative contribution of mutational signatures in each sample
}
\description{
Plotting function for mmsig package
}
