% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcheck.R
\name{refCheck}
\alias{refCheck}
\title{Check if the input mutation catalog is aligned to the correct reference genome}
\usage{
refCheck(snvs, genome)
}
\arguments{
\item{snvs}{input mutation catalog in 5-columns format}

\item{genome}{BSgenome object with the full reference genome used for analysis}
}
\value{
reference genome match: TRUE or FALSE
}
\description{
Check if the input mutation catalog is aligned to the correct reference genome
}
