% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat.R
\name{spat}
\alias{spat}
\title{Print message and variable}
\usage{
spat(dbg, name, var)
}
\arguments{
\item{dbg}{boolean whether to print or not}

\item{name}{message to print}

\item{var}{variable to print}
}
\value{
name and variable
}
\description{
Print message and variable
}
