% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spit.R
\name{spit}
\alias{spit}
\title{Print message}
\usage{
spit(dbg, mess, ...)
}
\arguments{
\item{dbg}{boolean whether to print or not}

\item{mess}{message to print}

\item{...}{additional parameters}
}
\value{
message
}
\description{
Print message
}
