% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfcleaner.R
\name{vcfCleaner}
\alias{vcfCleaner}
\title{Clean input VCF-like file}
\usage{
vcfCleaner(snvs)
}
\arguments{
\item{snvs}{input mutation catalog in 5-columns format}
}
\value{
clean mutation catalog in 5-columns format
}
\description{
Clean input VCF-like file
}
