% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandbias.R
\name{getStrandBias}
\alias{getStrandBias}
\title{Statistical tests for transcriptional strand bias}
\usage{
getStrandBias(data_5cols)
}
\arguments{
\item{data_5cols}{input mutational catalog in 5-column format}
}
\value{
list of results from strand bias in each trinucleotide context and pooled analysis for SBS-MM1 and SBS35
}
\description{
Statistical tests for transcriptional strand bias
}
